+++ evo1/checkpoints.py
@@ -1,10 +1,8 @@
-from evo1.seq.edel_vale import EdelExperimental
 from evo1.seq import OverworldToMeadow, MeadowFight
 
 
 def Checkpoints():
     return {
-        "edel_expr": EdelExperimental(),
         "overworld_expr": OverworldToMeadow(),
         "meadow_expr": MeadowFight(),
     }