+++ evo1/seq/__init__.py
@@ -1,8 +1,9 @@
-from evo1.seq.edel_vale import Edel1
-from evo1.seq.overworld import OverworldToMeadow, OverworldToCavern
+from evo1.seq.edel_vale import Edel1, Edel2
+from evo1.seq.overworld import OverworldToMeadow, OverworldToCavern, OverworldToNoria
 from evo1.seq.papurika import MeadowFight, PapurikaVillage
 from evo1.seq.start import Evoland1StartGame
 from evo1.seq.crystal_cavern import CrystalCavern
+from evo1.seq.noria_mines import NoriaMines
 
 __all__ = [
     "Evoland1StartGame",
@@ -12,4 +13,7 @@ __all__ = [
     "PapurikaVillage",
     "OverworldToCavern",
     "CrystalCavern",
+    "Edel2",
+    "OverworldToNoria",
+    "NoriaMines",
 ]