+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Current
 
+0.0.3:
+*
+
+# Older changelogs
+
 0.0.2:
 * TAS now loads maps from tmx (village/cavern)
 * TAS can now progress through cavern up to Kefka's Ghost (some pathing issues remain)
 * TAS can now kill the knights! No manual intervention required anymore
-
-# Older changelogs
 
 0.0.1:
 * Added timer to the main window